+++ src/main/java/pigcart/puddleflood/config/gui/ConfigResponders.java
@@ -1,50 +1,26 @@
 package pigcart.puddleflood.config.gui;
 
-import net.minecraft.ChatFormatting;
 import net.minecraft.client.Minecraft;
-import net.minecraft.network.chat.CommonComponents;
-import net.minecraft.network.chat.Component;
 
-import java.text.NumberFormat;
+import java.io.IOException;
 import java.util.function.Function;
+
+import static pigcart.puddleflood.VersionUtil.shadersEnabled;
 
 /// contains functions that are instantiated and run in response to config screen interactions
 public class ConfigResponders {
 
-    public static class Percent implements Function<Object, Component> {
-        public Component apply(Object value) {
-            return Component.literal(NumberFormat.getPercentInstance().format(value));
-        }
-    }
-
-    public static class PercentOrOff implements Function<Object, Component> {
-        public Component apply(Object value) {
-            return ((Number)value).floatValue() == 0 ? CommonComponents.OPTION_OFF.copy().withStyle(ChatFormatting.RED) : Component.literal(NumberFormat.getPercentInstance().format(value));
-        }
-    }
-
-    public static class ZeroIsAutomatic implements Function<Object, Component> {
-        public Component apply(Object stringValue) {
-            final int value = Integer.parseInt((String) stringValue);
-            return value == 0 ? Component.translatable("puddleflood.auto") : Component.literal((String) stringValue);
-        }
-    }
-
-    public static class ReloadResources implements Runnable {
+    public static class ReloadShaders implements Runnable {
         public void run() {
-            Minecraft.getInstance().reloadResourcePacks();
-        }
-    }
-
-    public static class ReloadChunks implements Runnable {
-        public void run() {
-            Minecraft.getInstance().levelRenderer.allChanged();
-        }
-    }
-
-    public static class RefreshScreen implements Runnable {
-        public void run() {
-            ((ConfigScreen)Minecraft.getInstance().screen).refresh();
+            //? <26.1 {
+            if (shadersEnabled()) {
+                try {
+                    net.irisshaders.iris.Iris.reload();
+                } catch (IOException e) {
+                    throw new RuntimeException(e);
+                }
+            }
+            //?}
         }
     }
 
@@ -55,13 +31,9 @@ public class ConfigResponders {
         }
     }
 
-    public static class IsNot1211 implements Function<Object, Boolean> {
+    public static class ShadersEnabled implements Function<Object, Boolean> {
         public Boolean apply(Object configContext) {
-            //? 1.21.1 {
-            /*return false;
-            *///?} else {
-            return true;
-            //?}
+            return shadersEnabled();
         }
     }
 }